HW_2, 1-3

+++ src/main/java/ru/javawebinar/topjava/repository/mock/mock/InMemoryMealRepositoryImpl.java
@@ -1,26 +1,31 @@
-package ru.javawebinar.topjava.repository.mock;
+package ru.javawebinar.topjava.repository.mock.mock;
 
+import ru.javawebinar.topjava.AuthorizedUser;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.util.exception.NotFoundException;
 
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 public class InMemoryMealRepositoryImpl implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
     {
-        MealsUtil.MEALS.forEach(this::save);
+        MealsUtil.MEALS.forEach(m -> save(m, AuthorizedUser.id()));
     }
 
     @Override
-    public Meal save(Meal meal) {
+    public Meal save(Meal meal, int userId) {
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
+            meal.setUserId(userId);
             repository.put(meal.getId(), meal);
             return meal;
         }
@@ -29,18 +34,32 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     }
 
     @Override
-    public void delete(int id) {
-        repository.remove(id);
+    public boolean delete(int id, int userId) {
+        if (repository.get(id) == null)
+            return false;
+
+        if (repository.get(id).getUserId() == userId) {
+            repository.remove(id);
+            return true;
+        } else
+            throw new NotFoundException("Wrong user");
     }
 
     @Override
-    public Meal get(int id) {
-        return repository.get(id);
+    public Meal get(int id, int userId) {
+        if (repository.get(id) == null)
+            return null;
+
+        if (repository.get(id).getUserId() == userId)
+            return repository.get(id);
+        else throw new NotFoundException("Its not your meal");
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public Collection<Meal> getAll(int userId) {
+        return repository.values().stream().filter(meal -> meal.getUserId() == userId).
+                sorted(Comparator.comparing(Meal::getDateTime).reversed())
+                .collect(Collectors.toList());
     }
 }
